Refactor SetOperator class into its own file
- syntax for (let element of someSet) { } did not work with keyword export; had to change
  method of iteration to someSet.forEach(element => { // DO SOMETHING });
